shodo-ecosystem/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
# uvloop==0.19.0  # removed for Python 3.13 compatibility
# httptools==0.6.1  # removed for Python 3.13 compatibility

# API Documentation
pydantic==2.10.6
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.37
alembic==1.13.1
# asyncpg==0.29.0  # removed for Python 3.13 compatibility (not needed for unit tests)
# psycopg2-binary==2.9.9  # removed for Python 3.13 compatibility (not needed for unit tests)

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Task Queue
celery==5.3.6
flower==2.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
orjson==3.9.10
email-validator==2.2.0

# Testing (dev only, but needed for CI)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0


# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.8.0
isort==5.13.2
bandit==1.7.6
tenacity==8.2.3
cachetools==5.3.2
psutil==5.9.6
